--- src/main/java/com/salttysugar/blog/storage/file/api/dto/FileDTO.java
@@ -1,17 +1,0 @@
-package com.salttysugar.blog.storage.file.api.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class FileDTO {
-    String id;
-    String name;
-    long size;
-    String url;
-}
