--- src/main/java/com/salttysugar/blog/storage/file/domain/model/storable/Storable.java
@@ -1,10 +1,0 @@
-package com.salttysugar.blog.storage.file.domain.model.storable;
-
-import reactor.core.publisher.Mono;
-
-import java.nio.file.Path;
-
-public interface Storable {
-    String getFileName();
-    Mono<Void> moveTo(Path path );
-}
