--- src/main/java/com/salttysugar/blog/storage/file/persistance/MongoFile.java
@@ -1,20 +1,0 @@
-package com.salttysugar.blog.storage.file.persistance;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-@Document
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Data
-public class MongoFile {
-    String id;
-    String name;
-    String path;
-    long size;
-    String type;
-}
