--- src/main/java/com/salttysugar/blog/storage/file/persistance/MongoFileRepository.java
@@ -1,6 +1,0 @@
-package com.salttysugar.blog.storage.file.persistance;
-
-import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
-
-public interface MongoFileRepository extends ReactiveMongoRepository<MongoFile, String> {
-}
