--- src/main/java/com/salttysugar/blog/storage/file/utils/converter/FileToFileDTOConverter.java
@@ -1,24 +1,0 @@
-package com.salttysugar.blog.storage.file.utils.converter;
-
-import com.salttysugar.blog.storage.common.constant.API;
-import com.salttysugar.blog.storage.file.api.dto.FileDTO;
-import com.salttysugar.blog.storage.file.domain.model.file.ApplicationFile;
-import lombok.SneakyThrows;
-import org.springframework.core.convert.converter.Converter;
-import org.springframework.stereotype.Component;
-
-@Component
-public class FileToFileDTOConverter implements Converter<ApplicationFile, FileDTO> {
-
-    @SneakyThrows
-    @Override
-    public FileDTO convert(ApplicationFile source) {
-        return FileDTO.builder()
-                .id(source.getId())
-                .name(source.getName())
-                .size(source.getSize())
-                //TODO: re-think it later
-                .url(String.format("%s/%s/meta", API.V1.FILE.BASE_URL, source.getId()))
-                .build();
-    }
-}
