--- src/main/java/com/salttysugar/blog/storage/file/utils/resolver/filetype/exception/UnknownFileTypeException.java
@@ -1,7 +1,0 @@
-package com.salttysugar.blog.storage.file.utils.resolver.filetype.exception;
-
-public class UnknownFileTypeException extends RuntimeException {
-    public UnknownFileTypeException() {
-        super("could not determine file extension");
-    }
-}
