--- src/main/java/com/salttysugar/blog/storage/file/utils/resolver/mediatype/MediaTypeResolver.java
@@ -1,10 +1,0 @@
-package com.salttysugar.blog.storage.file.utils.resolver.mediatype;
-
-import com.salttysugar.blog.storage.file.domain.constant.FileType;
-import org.springframework.http.MediaType;
-
-import java.util.Optional;
-
-public interface MediaTypeResolver {
-    Optional<MediaType> resolve(FileType type);
-}
