+++ src/main/java/com/salttysugar/blog/storage/model/Storable.java
@@ -1,0 +1,10 @@
+package com.salttysugar.blog.storage.model;
+
+import reactor.core.publisher.Mono;
+
+import java.nio.file.Path;
+
+public interface Storable {
+    String getFileName();
+    Mono<Void> moveTo(Path path );
+}
