+++ src/main/java/com/salttysugar/blog/storage/persistance/MongoFile.java
@@ -1,0 +1,20 @@
+package com.salttysugar.blog.storage.persistance;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+@Document
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Data
+public class MongoFile {
+    String id;
+    String name;
+    String path;
+    long size;
+    String type;
+}
