+++ src/main/java/com/salttysugar/blog/storage/persistance/MongoFileRepository.java
@@ -1,0 +1,6 @@
+package com.salttysugar.blog.storage.persistance;
+
+import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
+
+public interface MongoFileRepository extends ReactiveMongoRepository<MongoFile, String> {
+}
