+++ src/main/java/com/salttysugar/blog/storage/service/StorageService.java
@@ -1,0 +1,10 @@
+package com.salttysugar.blog.storage.service;
+
+import com.salttysugar.blog.storage.model.Storable;
+import reactor.core.publisher.Mono;
+
+import java.io.File;
+
+public interface StorageService {
+    Mono<File> store(Storable storable);
+}
