--- src/main/java/com/salttysugar/blog/storage/storage/config/StorageConfig.java
@@ -1,10 +1,0 @@
-package com.salttysugar.blog.storage.storage.config;
-
-import java.nio.file.Path;
-
-public interface StorageConfig {
-    Path getFolder();
-    default Path resolve(String filename) {
-        return getFolder().resolve(filename);
-    }
-}
