+++ src/main/java/com/salttysugar/blog/storage/utils/resolver/Functional.java
@@ -1,0 +1,14 @@
+package com.salttysugar.blog.storage.utils.resolver;
+
+import java.util.function.Consumer;
+import java.util.function.Function;
+import java.util.function.UnaryOperator;
+
+public final class Functional {
+    public static <T> Function<T,T> peek(Consumer<T> c) {
+        return x -> {
+            c.accept(x);
+            return x;
+        };
+    }
+}
