+++ src/main/java/com/salttysugar/blog/storage/utils/resolver/filetype/FileTypeResolver.java
@@ -1,0 +1,10 @@
+package com.salttysugar.blog.storage.utils.resolver.filetype;
+
+import com.salttysugar.blog.storage.model.FileType;
+import com.salttysugar.blog.storage.utils.resolver.filetype.exception.UnsupportedFileTypeException;
+
+import java.nio.file.Path;
+
+public interface FileTypeResolver {
+    FileType resolve(Path file) throws UnsupportedFileTypeException;
+}
