+++ src/main/java/com/salttysugar/blog/storage/utils/resolver/filetype/exception/UnknownFileTypeException.java
@@ -1,0 +1,7 @@
+package com.salttysugar.blog.storage.utils.resolver.filetype.exception;
+
+public class UnknownFileTypeException extends RuntimeException {
+    public UnknownFileTypeException() {
+        super("could not determine file extension");
+    }
+}
