+++ src/main/java/com/salttysugar/blog/storage/utils/resolver/filetype/exception/UnsupportedFileTypeException.java
@@ -1,0 +1,7 @@
+package com.salttysugar.blog.storage.utils.resolver.filetype.exception;
+
+public class UnsupportedFileTypeException extends RuntimeException {
+    public UnsupportedFileTypeException(String extension) {
+        super(String.format("unsupported file extension: %s", extension));
+    }
+}
