+++ src/main/java/com/salttysugar/blog/storage/utils/resolver/mediatype/MediaTypeResolver.java
@@ -1,0 +1,10 @@
+package com.salttysugar.blog.storage.utils.resolver.mediatype;
+
+import com.salttysugar.blog.storage.model.FileType;
+import org.springframework.http.MediaType;
+
+import java.util.Optional;
+
+public interface MediaTypeResolver {
+    Optional<MediaType> resolve(FileType type);
+}
